fix(alloc): correct type cast in String::pop
Changed type cast from u32 to usize when passing size to truncate().
This ensures correct sizing on 64-bit platforms and matches the
expected parameter type.

+++ src/sfc/alloc/string.cc
@@ -57,7 +57,7 @@ auto String::pop() noexcept -> Option<char32_t> {
   auto chars = this->as_str().chars();
 
   const auto ret = chars.next_back();
-  _vec.truncate(static_cast<u32>(chars._end - chars._ptr));
+  _vec.truncate(static_cast<usize>(chars._end - chars._ptr));
   return ret;
 }
